pyproject.toml
[tool.poetry]
name = "orderflow-v7"
version = "7.1.0"
description = "Rolling separability + adaptive TVTP + stability gates"
authors = ["Your Name <you@example.com>"]
license = "MIT"
package-mode = false

[tool.poetry.dependencies]
python = ">=3.10,<3.13"
pandas = ">=2.0.3,<3.0.0"
numpy = ">=1.24,<2.0"
scikit-learn = ">=1.3,<2.0"
hmmlearn = ">=0.3,<0.4"
statsmodels = ">=0.14,<0.15"
ccxt = ">=4.0,<5.0"
pyarrow = ">=14.0.1,<16.0.0"
pyyaml = ">=6.0,<7.0"
python-dotenv = ">=1.0,<2.0"
jsonschema = "^4.23.0"
tabulate = ">=0.9,<0.10"
polars = { version = ">=0.19.0,<0.21.0", optional = true }

[tool.poetry.group.dev.dependencies]
pytest = "^7"
black = "^23"
isort = "^5"
ruff = "^0.1"
mypy = "^1"

[tool.poetry.extras]
analytics = ["polars"]

[tool.ruff]
extend-select = ["TID251"]

[tool.mypy]
ignore_missing_imports = true

[tool.isort]
profile = "black"



[build-system]
requires = ["poetry-core>=1.0.0"]
build-backend = "poetry.core.masonry.api"
